split oneapi approx into approx1 and approx2 separate sources
